requirements.txt
# Ana Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Scanning & Security
python-nmap==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# Report Generation
jinja2==3.1.2
weasyprint==60.2  # PDF generation
matplotlib==3.8.2  # Graphs

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
click==8.1.7  # CLI framework